Move files, add jsconfig.json to client directory, update paths
