+++ server/server.js
@@ -5,9 +5,13 @@ const { readdir, readFile } = require('fs');
 const app = express();
 
 app.use(cors());
-app.use(express.static(path.resolve(__dirname, '../client', 'public')));
+app.use(express.static(path.resolve(__dirname, '../client', 'build')));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+
+app.get('*', (req, res) => {
+    res.sendFile(path.resolve(__dirname, '../client', 'build', 'index.html'));
+});
 
 app.get('/audio/:instrument', (req, res) => {
     const instrument = req.params.instrument;
